fix(function-team-player-count): correct returned table

+++ sqitch/deploy/function_team_player_count.sql
@@ -8,7 +8,7 @@
 BEGIN;
 
 CREATE FUNCTION trapparty.team_player_count(event_id INT)
-RETURNS TABLE(team_id INT, player_count INT) AS $$
+RETURNS TABLE(team_id INT, player_count BIGINT) AS $$
 BEGIN
   RETURN QUERY
     WITH data AS (SELECT player.id AS player_id, team.id AS team_id FROM trapparty.team, trapparty.player WHERE team.id = player.team_id AND team.event_id = 1) SELECT data.team_id, SUM(1) AS player_count FROM data GROUP BY data.team_id ORDER BY data.team_id ASC;
